+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
+import kitchenpos.domain.MenuProduct;
 
 public class Menu {
     private final Long id;
