+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao.repository;
+package kitchenpos.domain.menu;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -6,7 +6,6 @@ import java.util.Optional;
 import kitchenpos.dao.JdbcTemplateMenuDao;
 import kitchenpos.dao.JdbcTemplateMenuProductDao;
 import kitchenpos.dao.MenuDao;
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
 import org.springframework.stereotype.Repository;
