+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -17,7 +17,7 @@ import java.util.Optional;
 import java.util.stream.Stream;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
